Add files via upload
Problem Statement
Is there a relationship between humidity and temperature? What about between humidity and apparent temperature? Can you predict the apparent temperature given the humidity?

Statistics and Machine Learning can help us answer these questions and predict various relations and variables given the data.


Apparent temperature is the temperature that it feels like to our body because of other weather variables other than temperature. This is generally higher in case of high humidity.
In this project, we try to predict the apparent temperature given other weather variables in an hourly manner.
 

Objectives of this project

Clean the data and drop useless columns.
Make a EDA report .
Visualize the distributions of various features and correlations between them.